move clear_queue into queue
